Move LengthFormatter to strings module
