+++ audiorecorder/src/main/java/org/odk/collect/audiorecorder/recording/internal/AudioRecorderService.kt
@@ -8,7 +8,6 @@ import org.odk.collect.async.Scheduler
 import org.odk.collect.audiorecorder.getComponent
 import org.odk.collect.audiorecorder.recorder.Output
 import org.odk.collect.audiorecorder.recorder.Recorder
-import java.util.Locale
 import javax.inject.Inject
 
 class AudioRecorderService : Service() {
@@ -97,22 +96,5 @@ class AudioRecorderService : Service() {
 
         const val EXTRA_SESSION_ID = "EXTRA_SESSION_ID"
         const val EXTRA_OUTPUT = "EXTRA_OUTPUT"
-    }
-}
-
-object LengthFormatter {
-    const val ONE_HOUR = 3600000
-    const val ONE_MINUTE = 60000
-    const val ONE_SECOND = 1000
-
-    fun formatLength(milliseconds: Long): String {
-        val hours = milliseconds / ONE_HOUR
-        val minutes = milliseconds % ONE_HOUR / ONE_MINUTE
-        val seconds = milliseconds % ONE_MINUTE / ONE_SECOND
-        return if (milliseconds < ONE_HOUR) {
-            String.format(Locale.getDefault(), "%02d:%02d", minutes, seconds)
-        } else {
-            String.format(Locale.getDefault(), "%02d:%02d:%02d", hours, minutes, seconds)
-        }
     }
 }
